[PAD-49] MVP 배포

+++ src/main/java/com/example/padong_server/domain/subway/entity/Subway.java
@@ -31,11 +31,11 @@ public class Subway {
 
     // 출근 시간 혼잡도 합
     @Column(nullable = false)
-    private Integer morningCongestion;
+    private Double morningCongestion;
 
     // 퇴근 시간 혼잡도 합
     @Column(nullable = false)
-    private Integer eveningCongestion;
+    private Double eveningCongestion;
 
     // 위도
     @Column(nullable = false)
